Move create_person method to models
