+++ app/models/bgs_dependents/death.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module BGSDependents
+  class Death < Base
+    def initialize(death_info)
+      @death_info = death_info
+    end
+
+    def format_info
+      {
+        'death_date': @death_info['date'],
+        'vet_ind': 'N'
+      }.merge(@death_info['full_name'])
+    end
+  end
+end
